% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_name}
\alias{clean_name}
\title{clean_name}
\usage{
clean_name(name_string)
}
\arguments{
\item{name_string}{A character string from which to remove the code.}
}
\value{
The cleaned character string without the trailing code.
}
\description{
A helper function to remove trailing numeric codes from a string. It looks for
a dash followed by a space and one or more digits at the end of the string.
}
