% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAllConceptSets}
\alias{getAllConceptSets}
\title{getAllConceptSets}
\usage{
getAllConceptSets(con, cdmSchema)
}
\arguments{
\item{con}{A DBI connection to the database.}

\item{cdmSchema}{The name of the schema where the OMOP tables reside.}
}
\value{
A named list of `codelist` objects.
}
\description{
Retrieves all concept sets from the database and returns them as a named list of codelists.
}
