% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeCodelists}
\alias{mergeCodelists}
\title{mergeCodelists}
\usage{
mergeCodelists(..., newName)
}
\arguments{
\item{...}{An arbitrary number of `codelist` objects to be merged.}

\item{newName}{A character string specifying the name for the new merged codelist.}
}
\value{
A single, merged `codelist` object named according to the `newName` parameter.
}
\description{
Merges multiple `codelist` objects from the `omopgenerics` package into a
single new `codelist`. It combines all concept IDs from the input lists and
removes any duplicates.
}
\examples{
\dontrun{
  # Assuming bb_codes and ccb_codes are valid codelist objects
  first_line_therapies <- mergeCodelists(bb_codes, ccb_codes, newName = "first_line_hcm_drugs")
}
}
