% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_codelists}
\alias{process_codelists}
\title{process_codelists}
\usage{
process_codelists(codelist_vector)
}
\arguments{
\item{codelist_vector}{A list or vector containing the codelist objects, typically
created by a function like `getCodelistFromConceptSet()`.}
}
\value{
A new list with cleaned, descriptive names and containing only the first integer
vector of each original codelist.
}
\description{
Processes a vector of codelist objects. It iterates through the vector, extracts the internal,
descriptive name from each list's first element, removes any trailing numeric codes,
and then assigns the first integer vector to the new, cleaned name.
}
