% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCodelistFromConceptSet}
\alias{getCodelistFromConceptSet}
\title{getCodelistFromConceptSet}
\usage{
getCodelistFromConceptSet(conceptSetId, con, cdmSchema)
}
\arguments{
\item{conceptSetId}{The concept_set_id to query for.}

\item{con}{A DBI connection to the database.}

\item{cdmSchema}{The name of the schema where the OMOP tables reside.}
}
\value{
A named `codelist` object.
}
\description{
This function queries the `concept_set` and `concept_set_item` tables
to build a formal `codelist` object from the `omopgenerics` package.
The codelist is automatically named using the `concept_set_name` from the
`concept_set` table.
}
